tasks on more functions and more loopes

+++ 0x04-more_functions_nested_loops/2-mul.c
@@ -8,7 +8,7 @@
  */
 int mul(int a, int b)
 {
-	signed int i;
+	long i;
 
 	i = a * b;
 	return (i);
